YTDB-523: Unify duplicated extractKey() into parameterized helper
Both single-value and multi-value engines had identical extractKey()
methods differing only in the number of trailing elements to strip.
Extract a shared extractUserKey(CompositeKey, int) into VersionedIndexOps
and delegate from both engines.
